feat: #330 npm run format
