+++ packages/core/spec/assembly/filter/filter-set.spec.ts
@@ -1,9 +1,6 @@
 import { type PartsFilter } from '~core/assembly/filter/base'
 import { PartsFilterSet } from '~core/assembly/filter/filter-set'
-import {
-  enableOrNot,
-  type FilterType,
-} from '~core/assembly/filter/filter-type'
+import { enableOrNot, type FilterType } from '~core/assembly/filter/filter-type'
 
 import { fc, it } from '@fast-check/vitest'
 import sinon from 'sinon'
