+++ packages/core/spec/assembly/random/lock.spec.ts
@@ -5,7 +5,6 @@ import { boosterNotEquipped } from '@ac6_assemble_tool/parts/not-equipped'
 import { tank } from '@ac6_assemble_tool/parts/types/base/category'
 import { notEquipped } from '@ac6_assemble_tool/parts/types/base/classification'
 import { candidates } from '@ac6_assemble_tool/parts/versions/v1.06.1'
-
 import { it, fc, test } from '@fast-check/vitest'
 import { describe, expect } from 'vitest'
