+++ packages/core/spec/assembly/random/random-assembly.spec.ts
@@ -7,7 +7,6 @@ import type { Validator } from '~core/assembly/random/validator/base'
 import { failure, success } from '~core/assembly/random/validator/result'
 
 import { candidates } from '@ac6_assemble_tool/parts/versions/v1.06.1'
-
 import { fc, it } from '@fast-check/vitest'
 import { afterEach, beforeEach, describe, expect, type Mock, vi } from 'vitest'
