+++ packages/parts/src/booster.ts
@@ -1,6 +1,3 @@
-import type { BoosterNotEquipped } from '~data/not-equipped'
-import { boosterNotEquipped } from '~data/not-equipped'
-
 import { booster as boosterCategory } from './types/base/category'
 import { booster as boosterClass } from './types/base/classification'
 import {
@@ -11,6 +8,9 @@ import {
   schneider,
 } from './types/base/manufacture'
 import { defineBooster } from './types/inner/booster'
+
+import { boosterNotEquipped } from '~data/not-equipped'
+import type { BoosterNotEquipped } from '~data/not-equipped'
 
 export const boosters = [
   defineBooster({
