+++ packages/parts/src/expansions.ts
@@ -1,12 +1,12 @@
-import { expansion as expansionCategory } from '~data/types/base/category'
-import { expansion as expansionClass } from '~data/types/base/classification'
-import { noneManufacture } from '~data/types/base/manufacture'
-
 import {
   type AsAssaultArmor,
   type AsProtect,
   defineExpansion,
 } from './types/expansion/types'
+
+import { expansion as expansionCategory } from '~data/types/base/category'
+import { expansion as expansionClass } from '~data/types/base/classification'
+import { noneManufacture } from '~data/types/base/manufacture'
 
 export const expansions = [
   defineExpansion<AsAssaultArmor>({
