+++ packages/parts/src/versions/v1.07.ts
@@ -1,7 +1,7 @@
+import { candidates as v1_06_1, orders as order_v1_06_1 } from './v1.06.1'
+
 import type { Order } from '~data/types/candidates'
 import { apply } from '~data/versions/patches'
-
-import { candidates as v1_06_1, orders as order_v1_06_1 } from './v1.06.1'
 
 export const version = 'v1.07' as const
 export type VERSION = typeof version
