+++ packages/web/spec/error/error-message.spec.ts
@@ -1,3 +1,11 @@
+import type { I18Next } from '~view/i18n/define.ts'
+import {
+  assemblyErrorMessage,
+  filterApplyErrorMessage,
+  type Translator,
+} from '~view/pages/index/interaction/error-message'
+
+import { UsableItemNotFoundError } from '@ac6_assemble_tool/core/assembly/filter/filters'
 import { OverTryLimitError } from '@ac6_assemble_tool/core/assembly/random/random-assembly'
 import {
   notCarrySameUnitInSameSideName,
@@ -7,20 +15,10 @@ import {
   ValidationError,
   type ValidationName,
 } from '@ac6_assemble_tool/core/assembly/random/validator/validators'
-
-import type { I18Next } from '~view/i18n/define.ts'
-import {
-  assemblyErrorMessage,
-  filterApplyErrorMessage,
-  type Translator,
-} from '~view/pages/index/interaction/error-message'
-import { UsableItemNotFoundError } from '@ac6_assemble_tool/core/assembly/filter/filters'
 import { genAssemblyKey } from '@ac6_assemble_tool/core/spec-helper/property-generator'
-
 import { fc, it } from '@fast-check/vitest'
 import type { ArrayConstraints } from 'fast-check'
 import { afterEach, beforeEach, describe, expect, type Mock, vi } from 'vitest'
-
 
 describe(assemblyErrorMessage.name, () => {
   let i18n: Pick<I18Next, 't'>
