+++ packages/web/spec/pages/index/interaction/share.spec.ts
@@ -1,16 +1,14 @@
 import {
   assemblyToSearch,
   searchToAssembly,
-} from '~view/pages/index/interaction/share'
-
-import { it } from '@fast-check/vitest'
-import fc from 'fast-check'
-import { describe, expect } from 'vitest'
-
+} from '@ac6_assemble_tool/core/assembly/serialize/as-query'
 import {
   genAssembly,
   genCandidates,
-} from '~spec/spec-helper/property-generator'
+} from '@ac6_assemble_tool/core/spec-helper/property-generator'
+import { it } from '@fast-check/vitest'
+import fc from 'fast-check'
+import { describe, expect } from 'vitest'
 
 describe('query', () => {
   it.prop([
