+++ packages/web/src/pages/index/Index.svelte
@@ -16,7 +16,6 @@
     initialFilterState,
     toggleFilter,
   } from "~view/pages/index/interaction/filter.ts";
-  import { assemblyToSearch, searchToAssembly } from '~view/pages/index/interaction/share'
   import NavButton from "~view/pages/index/layout/navbar/NavButton.svelte";
   import ReportList from '~view/pages/index/report/ReportList.svelte'
   import ShareAssembly from '~view/pages/index/share/ShareAssembly.svelte'
@@ -32,6 +31,7 @@
   import {UsableItemNotFoundError} from "@ac6_assemble_tool/core/assembly/filter/filters";
   import {LockedParts} from "@ac6_assemble_tool/core/assembly/random/lock";
   import { RandomAssembly } from "@ac6_assemble_tool/core/assembly/random/random-assembly"
+  import {assemblyToSearch, searchToAssembly} from "@ac6_assemble_tool/core/assembly/serialize/as-query";
   import { logger } from '@ac6_assemble_tool/core/utils/logger'
   import {notEquipped} from "@ac6_assemble_tool/parts/types/base/category";
   import {type Candidates, defineOrder, type OrderParts} from "@ac6_assemble_tool/parts/types/candidates";
