+++ packages/web/src/pages/index/form/PartsSelectForm.svelte
@@ -1,7 +1,6 @@
 <script lang="ts" context="module">
   import type {AssemblyKey} from "@ac6_assemble_tool/core/assembly/assembly";
   import type {LockedParts} from "@ac6_assemble_tool/core/assembly/random/lock";
-
   import type {ACParts} from "@ac6_assemble_tool/parts/types/base/types.ts";
 
   export type ChangePartsEvent = Readonly<{ id: AssemblyKey, selected: ACParts }>
