+++ packages/web/tsconfig.json
@@ -17,5 +17,5 @@
     }
   },
   "include": ["./src/**/*", "./spec/**/*"],
-  "exclude": ["node_modules"],
+  "exclude": ["node_modules"]
 }
