+++ turbo.json
@@ -7,6 +7,12 @@
     "check-types": {
       "dependsOn": ["transit"]
     },
+    "lint": {
+      "dependsOn": ["transit"]
+    },
+    "format": {
+      "dependsOn": ["transit"]
+    },
     "dev": {
       "persistent": true,
       "cache": false
